cdk/pyproject.toml
[tool.poetry]
name = "cdk"
version = "0.1.0"
description = ""
authors = ["Martijn Sturm <martijn.sturm@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aws-cdk-lib = "^2.77.0"
constructs = "^10.2.10"
resource-registry = {path = "../../resource-registry", develop = true}
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
flake8 = "^6.0.0"
poetry = "^1.4.2"


[tool.poetry.group.test.dependencies]
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
